Fix small bug

+++ clean.sh
@@ -1,8 +1,16 @@
 #!/bin/bash
 SCRIPTPATH=$PWD
 
+if [ "$1" == "-soft" ]; then
+    SOFT="SET"
+fi
+
 for logFile in $(find $SCRIPTPATH -name "*.log" -o -name "*.cgmmOutput"); do
-    rm $logFile
+    if [ "$logFile" == "$SCRIPTPATH/Test_sup.log" ] && [ "$SOFT" ]; then
+        continue
+    else
+        rm $logFile
+    fi
 done
 
 for FDIR in $SCRIPTPATH/fingerprints/* ; do
@@ -11,16 +19,18 @@ for FDIR in $SCRIPTPATH/fingerprints/* ; do
     fi
 done
 
-for KDIR in $SCRIPTPATH/data/folder_* ; do
-    if [ -d $KDIR ]; then
-        rm -r $KDIR
-    fi
-done
-
 for LOGSFILE in $(ls $SCRIPTPATH/logsRun) ; do
     rm $SCRIPTPATH/logsRun/$LOGSFILE
 done
 
-if [ -d $SCRIPTPATH/data/KBase_folder ]; then
-    rm -r $SCRIPTPATH/data/KBase_folder
+if [ -z "$SOFT" ]; then
+    for KDIR in $SCRIPTPATH/data/folder_* ; do
+        if [ -d $KDIR ]; then
+            rm -r $KDIR
+        fi
+    done
+
+    if [ -d $SCRIPTPATH/data/KBase_folder ]; then
+        rm -r $SCRIPTPATH/data/KBase_folder
+    fi
 fi
